refactor(backend): consolidate orchestrator files — delete enhanced_orchestrator.py (#3393)
