+++ autobot-backend/services/advanced_workflow/step_generator.py
@@ -11,7 +11,7 @@ import logging
 from typing import List, Optional
 
 from autobot_types import TaskComplexity
-from enhanced_orchestrator import EnhancedOrchestrator, get_orchestrator
+from orchestrator import EnhancedOrchestrator, get_orchestrator_sync as get_orchestrator
 from type_defs.common import Metadata
 
 from .models import SmartWorkflowStep, WorkflowIntent
